+++ src/main.rs
@@ -1,6 +1,3 @@
-//! Substrate Node Template CLI library.
-#![warn(missing_docs)]
-
 mod chain_spec;
 mod rpc;
 #[macro_use]
